requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
sqlmodel[asyncio]==0.0.22
alembic==1.13.2
asyncpg==0.29.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
aiosqlite==0.20.0
email-validator==2.2.0
